requirements.txt
numpy>=1.25
scipy>=1.11
pandas>=2.0
pyyaml>=6.0
tqdm>=4.66
scikit-image>=0.22
scikit-learn>=1.3
laspy[lazrs]>=2.5
cloth-simulation-filter>=1.1
rasterio>=1.3
shapely>=2.0
fiona>=1.9
torch>=2.2
torchvision>=0.17
